refactor(VendingMachineManageMenu): 섹션 분리
